Fixed minor errors
