+++ core/prep_trim.py
@@ -4,7 +4,8 @@ import os
 import os.path
 # our modules
 from prep_trim_options import trim_illumina, trim_illumina_duplex
-from misc import trim_iontorrent
+sys.path.append(os.path.abspath(os.path.dirname(os.path.dirname(__file__))))
+from misc.process_ion import trim_iontorrent
 import primer_trim
 
 # Need to choose appropriate trimming function for different sequencing types
@@ -15,24 +16,6 @@ _trimming_function_ = {
 }
 
 #-------------------------------------------------------------------------------------
-def runShellCommand(cmd):
-    # run shell command, capture stdout and stderror (assumes log is not large and not redirected to disk)
-    try:
-        log = subprocess.check_output(cmd, shell=True, stderr=subprocess.STDOUT)
-        error = False
-    except subprocess.CalledProcessError, ex:
-        log = ex.output
-        error = True
-        
-    # print command log output
-    for line in log.split("\n"):
-        print("prep_trim: " + line.strip())
-     
-    # re-raise exception now that error detail has been printed
-    if error:
-        raise(ex)
-
-#-------------------------------------------------------------------------------------
 def run(cutadaptDir,tagNameUmiSeq,tagNamePrimer,tagNamePrimerErr,primer3Bases,filePrefix,primerFile,seqtype):
 
     print "Trimming synthetic regions , i.e. Adapters, UMI\n"
@@ -40,7 +23,7 @@ def run(cutadaptDir,tagNameUmiSeq,tagNamePrimer,tagNamePrimerErr,primer3Bases,fi
     print "Trimming Primers\n"
     fileOut1 = filePrefix + ".trimmed.R1.fastq"
     fileOut2 = filePrefix + ".trimmed.R2.fastq"   
-    primer_trim.main(fileIn1,fileIn2,fileOut1,fileOut2,primerFile,primerFile+'.clusters',int(primer3Bases),tagNamePrimer,tagNamePrimerErr,load_cache=True,cache_file=primerFile+".kmer.cache")
+    primer_trim.main(fileIn1,fileIn2,fileOut1,fileOut2,primerFile,primerFile+'.clusters',int(primer3Bases),tagNamePrimer,tagNamePrimerErr,load_cache=False,cache_file=primerFile+".kmer.cache")
     print "Finished with trimming\n"    
     # delete unneeded temp files
     for i in range(0,3):
@@ -57,9 +40,8 @@ if __name__ == "__main__":
     tagNameUmiSeq = sys.argv[2]
     tagNamePrimer = sys.argv[3]
     tagNamePrimerErr = sys.argv[4]
-    primerFasta = sys.argv[5]
-    primer3Bases = sys.argv[6]
-    readFilePrefix = sys.argv[7]
-    primerFile = sys.argv[8]
-    seqtype = sys.argv[9]
-    run(cutadaptDir,tagNameUmiSeq,tagNamePrimer,tagNamePrimerErr,primerFasta,primer3Bases,readFilePrefix,primerFile,seqtype)
+    primer3Bases = sys.argv[5]
+    readFilePrefix = sys.argv[6]
+    primerFile = sys.argv[7]
+    seqtype = sys.argv[8]
+    run(cutadaptDir,tagNameUmiSeq,tagNamePrimer,tagNamePrimerErr,primer3Bases,readFilePrefix,primerFile,seqtype)
